Misc enhancements to the broker routing
- Replace the buffer with wrapped buffer to save the bytes copy
- Change the response size and delay return type to int

+++ pinot-core/src/main/java/org/apache/pinot/core/transport/QueryRouter.java
@@ -104,8 +104,8 @@ public class QueryRouter {
     _serverChannels.shutDown();
   }
 
-  void receiveDataTable(ServerRoutingInstance serverRoutingInstance, DataTable dataTable, long responseSize,
-      long deserializationTimeMs) {
+  void receiveDataTable(ServerRoutingInstance serverRoutingInstance, DataTable dataTable, int responseSize,
+      int deserializationTimeMs) {
     long requestId = Long.parseLong(dataTable.getMetadata().get(DataTable.REQUEST_ID_METADATA_KEY));
     AsyncQueryResponse asyncQueryResponse = _asyncQueryResponseMap.get(requestId);
